% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcTimeAllCpp}
\alias{CalcTimeAllCpp}
\title{Multiply a number by two}
\usage{
CalcTimeAllCpp(ID, Abs_dates, Rating, labels, end_date)
}
\arguments{
\item{x}{A single integer.}
}
\description{
Multiply a number by two
}

