% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transprob.R
\name{DegradeOnly}
\alias{DegradeOnly}
\title{Transforms estimated transition matrices into upper diagonal matrices in a crude way}
\usage{
DegradeOnly(transmatrix, LoadDiag = TRUE)
}
\arguments{
\item{transmatrix}{an estimated transition matrix from transprob}

\item{LoadDiag}{logical. If TRUE (default) lower diagonal elements are added onto the diagonal element in the same row. 
If FALSE, lower diagonal elements are set to zero and each row rescaled to ensure a row sum of 1.}
}
\description{
Transforms estimate transition matrices to be upper diagonal either by adding all lower diagonal elements onto the diagonal 
of the same row, or by setting lower diagonal elements to 0 and rescaling the other row elements so that the row sum remains 1.
}

