% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Embankment}
\alias{Embankment}
\title{Embankment data}
\format{A data frame with 178515 rows and 3 columns:
\describe{
  \item{ID}{Unique asset ID}
  \item{Date}{Inspection date in format \%d-\%b-\%Y}
  \item{EHC}{Earthworks hazard category: a factor with value in 'A', 'B','C','D', 'E'}
}}
\usage{
Embankment
}
\description{
A dataset containing 178515 ratings of 97076 unique embankment assets
}
\keyword{datasets}

