% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FinancialRating}
\alias{FinancialRating}
\title{Financial ratings of 1505 financial assets}
\format{A data frame with 4315 rows and 3 variables:
\describe{
  \item{ID}{Unique asset ID}
  \item{Date}{Inspection date in format \%d-\%b-\%Y}
  \item{Rating}{Financial rating: a factor with value in 'A', 'AA', 'AAA', 'B','BB','BBB','CCC','D'}
}}
\source{
MATLAB; load Data_TransProb
}
\usage{
FinancialRating
}
\description{
A dataset containing 4315 ratings of 1505 financial assets.
}
\keyword{datasets}

